+++ src/evaluate.js
@@ -1,47 +1,67 @@
 const stdlib = require("./stdlib");
+const _Boolean = require("./stdlib/types/Boolean");
 const {
   createEnv,
   setEnv,
-  lookup,
   getValue,
-  setValue,
   defVar,
-  getIdentifier,
 } = require("./environment");
+const { ArithTypeError } = require("./errors");
+const environment = setEnv(stdlib);
+
 const { tokenize } = require("./tokenize");
 const { parse } = require("./parse");
 
-const environment = setEnv(stdlib);
-
-const define = (node, env = environment) => {
-  return defVar(node.name, evaluate(node.value, env), env);
+const evaluate = (node, env = environment) => {
+  switch (node.type) {
+    case "Program":
+      return evalBlock(node.body);
+    case "DecimalLiteral":
+      return stdlib.decimal(node.value);
+    case "StringLiteral":
+      return stdlib.string(node.value);
+    case "NilLiteral":
+      return stdlib.nil;
+    case "BooleanLiteral":
+      return new _Boolean(node.value);
+    case "Identifier":
+      return getValue(node, env);
+    case "CallExpression":
+      return apply(node, env);
+    case "DefinitionExpression":
+      return define(node, env);
+    case "LambdaExpression":
+      return makeLambda(node, env);
+    case "IfExpression":
+      return applyIf(node, env);
+  }
 };
 
-const apply = (node, env = environment) => {
+const evalBlock = (block, env) => {
+  let val;
+  for (let i = 0; i < block.length; i++) {
+    val = evaluate(block[i], env);
+  }
+  return val;
+};
+
+const apply = (node, env) => {
   const fn = getValue(node, env);
   const name = fn.name || node.name;
-
-  const args = node.arguments.map((a, i) => evaluate(a, env));
-  if (typeof fn !== "function") {
-    throw new TypeError(`${fn} is not a function`);
+  const args = node.arguments.map((a) => {
+    return evaluate(a, env);
+  });
+  if (fn instanceof Function !== true) {
+    throw new ArithTypeError(`${name} is not a function`);
   }
-
   return fn(...args);
 };
 
-const applyKeyword = (node, env = environment) => {
-  const name = `${node.name}Expr`;
-
-  return apply(
-    {
-      ...node,
-      name,
-    },
-    env,
-  );
+const define = (node, env) => {
+  return defVar(node.name, evaluate(node.value, env), env);
 };
 
-const makeLambda = (node, env = environment) => {
+const makeLambda = (node, env) => {
   const lambda = (...args) => {
     const names = node.params;
     const scope = createEnv(env);
@@ -50,66 +70,17 @@ const makeLambda = (node, env = environment) => {
         defVar(n.name, args[i], scope);
       });
     }
-    return evaluate(node.body, scope);
+    return evalBlock(node.body, scope);
   };
   return lambda;
 };
 
-const applyIf = (node, env = environment) => {
-  const cond = evaluate(node.condition, env);
-
+const applyIf = (node, env) => {
+  const cond = evaluate(node.condition);
   if (cond !== false) {
     return evaluate(node.then, env);
-  } else {
-    return evaluate(node.else, env);
   }
+  return evaluate(node.else, env);
 };
 
-const evaluate = (node, env = environment) => {
-  switch (node.type) {
-    case "Identifier":
-      return getValue(node, env);
-
-    case "KeywordExpression":
-      return applyKeyword(node, env);
-
-    case "CallExpression":
-      return apply(node, env);
-
-    case "DefinitionExpression":
-      return define(node, env);
-
-    case "LambdaExpression":
-      return makeLambda(node, env);
-
-    case "IfExpression":
-      return applyIf(node, env);
-  }
-
-  if (node.value) {
-    return node.value;
-  } else if (node.value === 0) {
-    return 0;
-  } else if (node.value === false) {
-    return false;
-  } else if (node.value === "") {
-    return "";
-  } else {
-    throw new TypeError(`${node.type} is invalid`);
-  }
-};
-
-const evaluateProgram = (prog) => {
-  let i = 0;
-  let val;
-  while (i < prog.body.length) {
-    if (prog.body[i]) {
-      val = evaluate(prog.body[i]);
-    }
-    i += 1;
-  }
-
-  return val;
-};
-
-module.exports = { evaluate, evaluateProgram };
+module.exports = { evaluate };
